+++ backend/django_project/urls.py
@@ -4,9 +4,9 @@ The `urlpatterns` list routes URLs to views. For more information please see:
     https://docs.djangoproject.com/en/4.1/topics/http/urls/
 """
 from django.conf import settings
+from django.conf.urls.static import static
 from django.contrib import admin
 from django.urls import include, path
-from django.conf.urls.static import static
 from django.utils.translation import gettext_lazy as _
 
 urlpatterns = [